requirements.txt
# 核心依赖
qiskit
hydra-core>=1.3
omegaconf>=2.3
pymoo>=0.6
pyscf>=2.4
pytest-cov>=4.1
qiskit-aer>=0.13
qiskit-nature>=0.7
scienceplots>=2.1
wandb>=0.16

# PyTorch Geometric 相关（需要与已安装的 PyTorch 版本兼容）
torch-geometric>=2.4
# torch-scatter 和 torch-sparse 的版本需与 PyTorch 和 CUDA 匹配，建议单独安装
# 详见下方说明
